Updates for new metadata_json format and other minor fixes (#33)
* Refine bio script modifications
* Edits to parseRefineBioMetadata to use platformAccessionCode as the platform id
* Various updates, const, atomic, small script improvements
* Update instructions for SeekEvaluator

+++ scripts/seek/build_species_compendium.sh
@@ -3,26 +3,42 @@
 ########
 ### Script to run the various steps to build a species Seek compendium
 #
+# Required inputs: pcl files, quant2 file refine_bio_metadata.json file
+#
+# Additionally, the refinebio_platforms will be downloaded from
+# https://api.refine.bio/v1/platforms/
+#
 # Edit the first 5 lines below for the species name, NCBI location and
 #   base diretory information
 #
 # Typical species built:
-#   yeast : saccharomyces_cerevisiae
-#   mouse : mus_musculus
-#   worm : caenorhabditis_elegans
 #   fly : drosophila_melanogaster
+#       NCBI_GENE_INFO='https://ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/Invertebrates/Drosophila_melanogaster.gene_info.gz'
+#   human : homo_sapiens
+#       NCBI_GENE_INFO='https://ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/Mammalia/Homo_sapiens.gene_info.gz'
+#   mouse : mus_musculus
+#       NCBI_GENE_INFO='https://ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/Mammalia/Mus_musculus.gene_info.gz'
+#   worm : caenorhabditis_elegans
+#       NCBI_GENE_INFO='https://ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/Invertebrates/Caenorhabditis_elegans.gene_info.gz'
+#   yeast : saccharomyces_cerevisiae
+#       NCBI_GENE_INFO='https://ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/Fungi/Saccharomyces_cerevisiae.gene_info.gz'
 #   zebrafish : danio_rerio
+#       NCBI_GENE_INFO='https://ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/Non-mammalian_vertebrates/Danio_rerio.gene_info.gz'
 ########
 
 SPECIES_NAME='saccharomyces_cerevisiae'
 SHORT_NAME='yeast'
 NCBI_GENE_INFO='https://ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/Fungi/Saccharomyces_cerevisiae.gene_info.gz'
+
 OUTPUT_BASE_DIR='/data/seek/tmp'
 OUTPUT_DIR="${OUTPUT_BASE_DIR}/${SHORT_NAME}"
-
-NEW_PCL_DATA='/Genomics/function/pentacon/akwong/projects/modSeek_rpc/refine_bio_nonhuman/'
+NEW_PCL_DATA='/Genomics/function/pentacon/akwong/projects/modSeek_rpc/compendia/pcls'
 SCRIPTS_DIR=$(realpath ~/src/github/FunctionLab/sleipnir/scripts/seek)
-SEEK_BIN=$(realpath ~/src/github/FunctionLab/sleipnir/Debug)
+SEEK_BIN=$(realpath ~/src/github/FunctionLab/sleipnir/Release)
+
+REFINE_BIO_FILE=$(realpath ${OUTPUT_BASE_DIR}/refine_bio_meta.json)
+REFINE_PLATFORMS_FILE=$(realpath ${OUTPUT_BASE_DIR}/refine_bio_platforms.json)
+QUANT_FILE=$(realpath ${OUTPUT_BASE_DIR}/quant2)
 
 # Activate conda environment
 if [ -z $CONDA_DEFAULT_ENV ] || [ $CONDA_DEFAULT_ENV != "genomics" ]; then
@@ -37,7 +53,24 @@ if [ -d ${OUTPUT_DIR} ]; then
 fi
 mkdir ${OUTPUT_DIR}
 cd ${OUTPUT_DIR}
-cp ../quant2 .
+
+# check if required files exist
+if [ ! -f ${REFINE_BIO_FILE} ]; then
+    echo "Expecting file refine_bio_meta.json " \
+    "in output base directory ${OUTPUT_BASE_DIR}"
+    exit -1
+fi
+if [ ! -f ${REFINE_PLATFORMS_FILE} ]; then
+    # Download the platform metadata from https://api.refine.bio/v1/platforms/
+    wget https://api.refine.bio/v1/platforms/ -O ${REFINE_PLATFORMS_FILE}
+fi
+if [ ! -f ${QUANT_FILE} ]; then
+    echo "Expecting file quant2 " \
+    "in output base directory ${OUTPUT_BASE_DIR}"
+    exit -1
+fi
+
+cp ${QUANT_FILE} .
 
 # 2. Get the gene info
 wget ${NCBI_GENE_INFO} -O ${SPECIES_NAME}.gene_info.gz
@@ -53,7 +86,8 @@ grep -e protein-coding -e rRNA gene_types_map.txt | cut -f 1 > coding_gene_list.
 # 3. Parse refine_bio info and make the dsetPlatMap file
 # Note: first command below outputs file dset_map.txt
 python ${SCRIPTS_DIR}/parseRefineBioMetadata.py  \
-    -f ../refine_bio_nonhuman_meta.json \
+    -f ${REFINE_BIO_FILE} \
+    -p ${REFINE_PLATFORMS_FILE} \
     -s ${SPECIES_NAME} \
     --make-dset-map -o .
 cut -f 3,4 dset_map.txt > dsetPlatMap.txt
